Renamed why-divest dirs to fossil-fuel for cont.
