+++ src/components/navigation/PrimaryNavBar.tsx
@@ -49,7 +49,7 @@ const PrimaryNavBar: FC = () => {
         <Navbar.Collapse>
           <NavButton title="Home" link="/home" />
           <NavButton title="Take Action" link="/take-action" />
-          <NavButton title="Why Divest?" link="/why-divest" />
+          <NavButton title="Why Divest?" link="/fossil-fuel" />
           <NavButton title="Investment Database" link="/investments" />
           <div className="hidden w-fit flex-col justify-center md:flex">
             <AuthButton />
