+++ src/pages/fossil-fuel/index.tsx
