+++ src/sections/fossil-fuel/DirtyIndustries.tsx
