+++ src/sections/index.tsx
@@ -10,13 +10,13 @@ import TimelineSection from './home/Timeline'
 import ChartGroup from './company/ChartGroup'
 import CompanyTooltipGroup from './company/TooltipGroup'
 //Fossil Fuel Page
-import DivestmentCase from './why-divest/DivestmentCase'
-import DirtyIndustries from './why-divest/DirtyIndustries'
-import FossilFuelsBad from './why-divest/FossilFuelsBad'
-import InstitutionsDivested from './why-divest/InstitutionsDivested'
-import SchoolsDivested from './why-divest/SchoolsDivested'
-import UniversityInvestments from './why-divest/UniversityInvestments'
-import WhatWarmingMeans from './why-divest/WhatWarmingMeans'
+import DivestmentCase from './fossil-fuel/DivestmentCase'
+import DirtyIndustries from './fossil-fuel/DirtyIndustries'
+import FossilFuelsBad from './fossil-fuel/FossilFuelsBad'
+import InstitutionsDivested from './fossil-fuel/InstitutionsDivested'
+import SchoolsDivested from './fossil-fuel/SchoolsDivested'
+import UniversityInvestments from './fossil-fuel/UniversityInvestments'
+import WhatWarmingMeans from './fossil-fuel/WhatWarmingMeans'
 // Take Action Page
 import OpenLetter from './take-action/letter'
 import SignLetter from './take-action/sign-letter'
